Rework spawn world selection into it's own event.
The player is created a bit before we call the PlayerJoinEvent, and by
that time the player has already ticked and been sent some chunks for
their current world, which the PlayerJoinEvent wouldn't have had a
chance to modify.

This introduces a PlayerSpawnLocationEvent, which allows a spawn
location to be chosen for a player right as they are created, before
PlayerJoinEvent is called.

This solution is pretty bad. If you have any better way to implement
this correctly then certainly tell.