This commit fixes https://github.com/LOOHP/Limbo/issues/51
"Incorrect Payload Received in KeepAlive packet for player"

The problem is that in my setup (described in the issue linked above)
the loohp-limbo server is receiving an unsolicited PacketInKeepAlive
message prior to sending any out.  This causes the payload validation
check to fail, resulting in the connection then being closed.

This commit changes the PacketInKeepAlive handler to ignore (other than
logging) unsolicited KeepAlive messages rather than act on them.

Possibly the logging is a bad idea unless it's limited to being
suppressed via a debug flag...